# sql/tests: use SFU with consistent lock ordering in Bank benchmark